Decrypter melhorado
Codigo colocando dentro de uma função, melhor organizado, (with) usado para o proprio python fechar o arquivo e nao der nenhum erro na hora da descriptografia, (try) e (except) usados caso acontecer algum erro no arquivo ou corrompimento.

+++ decrypter.py
@@ -1,22 +1,30 @@
 import os
 import pyaes
 
-## abrir o arquivo criptografado
-file_name = "teste.txt.ransomwaretroll"
-file = open(file_name, "rb")
-file_data = file.read()
-file.close()
+def descriptografar():
+  try:
+    file_name= "teste.txt.ransomwaretroll"
+    key = b"testeransomwares"
 
-## chave para descriptografia
-key = b"testeransomwares"
-aes = pyaes.AESModeOfOperationCTR(key)
-decrypt_data = aes.decrypt(file_data)
+    if not os.path.exists(file_name):
+      print(f" -- Erro: O arquivo {file_name} não foi encontrado.")
+      return 
+    with open(file_name, "rb") as file:
+      file_data = file.read()
 
-## remover o arquivo criptografado
-os.remove(file_name)
+        aes = pyaes.AESModeOfOperationCTR(key)
+        decrypt_data = aes.decrypt(file_data)
 
-## criar o arquivo descriptografado
-new_file = "teste.txt"
-new_file = open(f'{new_file}', "wb")
-new_file.write(decrypt_data)
-new_file.close()
+        os.remove(file_name)
+      
+        new_file_name = "teste.txt"
+        with open(new_file_name, "wb") as new_file:
+            new_file.write(decrypt_data)
+
+        print(f"-- Arquivo '{file_name}' recuperado com sucesso para '{new_file_name}'")
+
+    except Exception as e:
+        print(f"-- Erro na descriptografia: {e}")
+
+if __name__ == "__main__":
+    descriptografar()
